refactor deploy into multiple files
