+++ Cargo.toml
@@ -15,3 +15,6 @@ mime_guess = "2.0.3"
 multipart = { version = "0.18.0", features = ["client"] }
 sha2 = "0.9.8"
 difference = "2.0.0"
+rusoto_core = "0.47.0"
+rusoto_s3 = "0.47.0"
+tokio = { version = "1", features = ["full"] }
